feat: implement database and analytics API integration for 100% completion
- Add comprehensive database API proxy routes (/api/database/connections, /api/database/tables, /api/database/query, /api/database/vector-search)
- Implement safety constraints: SQL injection protection, timeout limits, row limits for database operations
- Add analytics API proxy routes (/api/analytics) with anomaly detection, trends, and forecasting
- Update database API client to use real V3 backend endpoints instead of mock data
- Update analytics API client with real backend integration for summary, anomalies, and predictions
- Fix syntax errors and clean up database API implementation
- Update dashboard status to reflect database and analytics integrations as partially complete

BREAKING: Database and analytics dashboards now connect to real V3 backend APIs instead of mock data, enabling safe SQL execution, vector search, and statistical analysis

+++ iterations/v3/indexers/src/database.rs
@@ -44,7 +44,7 @@ impl DatabasePool {
 
     /// Get pool size
     pub fn size(&self) -> usize {
-        self.pool.size()
+        self.pool.size() as usize
     }
 }
